% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossing.probability.R
\name{crossing.prob}
\alias{crossing.prob}
\title{Crossing Probability}
\usage{
crossing.prob(
  lower.boundaries = NULL,
  upper.boundaries = NULL,
  method = c("auto", "ecdf1-mns2016", "ecdf1-new", "ecdf2-ks2001", "ecdf2-mn2017")
)
}
\arguments{
\item{lower.boundaries}{Lower boundaries. Defaults to \code{NULL}.}

\item{upper.boundaries}{Upper boundaries. Defaults to \code{NULL}.}

\item{method}{See Details. Defaults to \code{"auto"}.}
}
\value{
A float giving probability of Poisson process to stay
within the boundaries.
}
\description{
Calculates the boundary crossing probability in either a one- or
two-sided case.
\loadmathjax
}
\details{
For \mjeqn{n}{n} i.i.d. uniformly distributed random variables
\mjeqn{X_1, ..., X_n}{X1, ..., Xn} \mjeqn{X_i \sim U(0,1)}{Xi ~ U(0,1)}
the empirical cumulative distribution function is:

\mjdeqn{\hat{F}_n(t) = \frac{1}{n}\sum_i{\mathbb{1}(
X_i \leq t)}.}{F^(t)_n = 1/n sum_i(1(Xi <= t)).
}

For two monotone-increasing, non-crossing functions \mjeqn{g,h : [0,1]
\rightarrow \mathbb{R}}{g,h : [0,1] -> R} function the
two-sided non-crossing probability is:

\mjdeqn{P(\forall t \in [0, 1] : g(t) < \hat{F}_n(t) < h(t) )
.}{P(forall t in [0,1] : g(t) < F^(t)_n < h(t) ).}

See \insertCite{NEWEST;textual}{crossingprobability} for
more details.

We provide the following methods as implemented by Amit Moscovich:

\itemize{
\item{\code{"ecdf1-mns2016"}}{\mjeqn{O(n^2)}{O(n^2)}
\insertCite{MNS2016}{crossingprobability}}
\item{\code{"ecdf1-new"}}{\mjeqn{O(n^2)}{O(n^2)}
\insertCite{NEWEST}{crossingprobability}}
\item{\code{"ecdf2-ks2001"}}{\mjeqn{O(n^3)}{O(n^3)}
\insertCite{KS2001}{crossingprobability}}
\item{\code{"ecdf2-mn2017"}}{\mjeqn{O(n^2 \log{n})}{O(n^2 log(n))}
\insertCite{MN2017}{crossingprobability}}
}

The default \code{"auto"} selects \code{"ecdf1-new"} for one-sided
boundaries and \code{"ecdf2-mn2017"} for two-sided uses. Note that
the crossing probability is returned and not the not crossing probability.
}
\examples{
b <- c(0.1, 0.1, 0.3)
B <- c(0.25, 0.5, 0.7)
crossing.prob(b, B)
}
\references{
\insertAllCited{}
}
